requirements.txt
twitchio
python-telegram-bot